Desktop/first mini app telgram/44444444444/kyc-verification-app-master/kyc-verification-app-master/requirements.txt
Flask>=2.2.5
Flask-Cors>=4.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
Flask-SQLAlchemy>=3.1.1
psycopg2-binary>=2.9.1
python-telegram-bot>=20.7
nest-asyncio>=1.5.8
requests>=2.26.0
pyngrok>=7.0.0
